:truck: re-organize mdx components
