+++ src/components/MarkdownContent/MarkdownContent.js
@@ -2,29 +2,8 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import { MDXProvider } from '@mdx-js/react'
 import { Box } from 'rebass'
-import * as Heading from './Headings'
-import { UnorderedList, OrderedList } from './Lists'
-import { Paragraph, BlockQuote, HorizontalRule } from './TypographicElements'
-import { Code, InlineCode } from './Code'
-import { Image } from './Embeds'
+import components from './components'
 import { verticalRhythm } from './markdownStyles'
-
-const components = {
-  h1: Heading.One,
-  h2: Heading.Two,
-  h3: Heading.Three,
-  h4: Heading.Four,
-  h5: Heading.Five,
-  h6: Heading.Six,
-  ul: UnorderedList,
-  ol: OrderedList,
-  p: Paragraph,
-  blockquote: BlockQuote,
-  hr: HorizontalRule,
-  code: Code,
-  inlineCode: InlineCode,
-  img: Image,
-}
 
 const MarkdownContent = ({ children, ...props }) => (
   <MDXProvider components={components}>
