+++ src/components/MarkdownContent/components/Headings.js
@@ -2,7 +2,7 @@ import React from 'react'
 import PropTypes from 'prop-types'
 import GithubSlugger from 'github-slugger'
 import { Link } from 'rebass'
-import { Heading } from '../Typography'
+import { Heading } from '../../Typography'
 
 const LinkIcon = () => (
   <svg
